kafka_producer/requirements-producer.txt
pandas==2.2.2
numpy==1.26.4
pydantic==2.7.4
kafka-python==2.0.2
websockets==12.0
pyarrow==16.1.0